Active gantt scroll bar, fix tabs and address pr feedback
